planta buena

+++ src/programaciondmi/dca/ecosistemas/piedrahitarinconrojasrojas/PlantaMala.java
@@ -98,7 +98,7 @@ public class PlantaMala extends PlantaAbstracta {
 		}
 	}
 
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
@@ -106,7 +106,7 @@ public class PlantaMala extends PlantaAbstracta {
 		this.x = x;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
